delete friends from other user's friends lists when user is deleted

+++ server/src/services/userService.js
@@ -102,10 +102,18 @@ async function updateProfile(user, password) {
 
 // delete user by id
 async function deleteUserById(userId) {
-    if(await userDAO.deleteUserById(userId)) {
-        return true;
-    }
-    return false;
+  const user = await userDAO.findUserById(userId);
+  if(!user) return false;
+
+  const friends = user.friends;
+  for(let friendId of friends) {
+    await removeFriend(friendId, userId);
+  }
+
+  if(await userDAO.deleteUserById(userId)) {
+    return true;
+  }
+  return false;
 }
 
 // find user by id
@@ -148,7 +156,17 @@ async function getUsersFriends(user) {
     };
 }
 
+async function removeFriend(userIdToRemoveFrom, friendId) {
+  const userToUpdate = await userDAO.findUserById(userIdToRemoveFrom);
+  userToUpdate.friends = userToUpdate.friends.filter((element) => element != friendId);
+  if(!await userDAO.updateUser(userToUpdate)) {
+    return false;
+  }
+  return true;
+}
+
+
 module.exports = {
-  registerUser, loginUser, getStats, updateProfile, deleteUserById, findUserById, getUsersFriends, updateAccount
+  registerUser, loginUser, getStats, updateProfile, deleteUserById, findUserById, getUsersFriends, removeFriend, updateAccount
 };
